frick you
sdjdkxjfhkdsjghdj

+++ equation_solution.h
@@ -7,7 +7,7 @@
 
 
 int square_equation_solution(double a, double b, double c, double* x1, double* x2, char* err);
-double linear_equation_solution(double b, double c, double* x1, char*err);
+int linear_equation_solution(double b, double c, double* x1, char*err);
 double discriminant(double a, double b, double c);
